// src/models/Member.ts
export default class Member {
  id: number = 0;
  image?: string = "";
  name?: string = "";
  type?: string = "";
  number?: string = "";
  position?: string = "";
  throwing?: string = "";
  hitting: string = "";
  avg: string = "0";
  era: string = "0";
  order_id?: string | number = 0;
  order?: string | number = "";
}